Начал работать над выводом в консоль
